requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
langchain==0.0.314
langchain-google-genai==0.0.5
langchain-huggingface==0.0.7
pydriller==2.5.0
gitpython==3.1.40
faiss-cpu==1.7.4
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.34.1
torch==2.1.0